Refactored for maven and some other improvements
